+++ neural_net.h
@@ -43,7 +43,7 @@ Matrix *scalar_p_M(Matrix *A, double n);
 Matrix *transpose_M(Matrix *A);
 void sum_M(Matrix *Res, Matrix *A);
 void print_M(Matrix *A);
-void fill_from_array_M(Matrix *A, int8_t *arr, unsigned int len);
+void fill_from_array_M(Matrix *A, double *arr, unsigned int len);
 void rand_M(Matrix *A, double min, double max);
 void ascii_print_M(Matrix *A);
 MLP *newMLP(int depth, int input_size, int output_size, int hidden_layer_size, activation_f* activations);
